requirements.txt
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.12.2
reportlab==4.0.5
python-docx==0.8.11
SQLAlchemy==2.0.20
schedule==1.2.0
pyyaml==6.0.1
